mappings and activity moved to new design pattern
